jspAction 수정

+++ jspAction/scheduleDeleteAction.jsp
@@ -9,16 +9,31 @@
 
     int scheduleIdxValue = Integer.parseInt(request.getParameter("scheduleIdx"));
     String dateValue = request.getParameter("date");
+    String errMessage = "";
 
-    Class.forName("com.mysql.jdbc.Driver"); 
-    Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/scheduler", "suin", "suin"); 
+    boolean isLoginIn = true; 
 
-    String sql = "DELETE FROM schedule WHERE idx=? AND user_idx=?";
-    PreparedStatement query = conn.prepareStatement(sql); 
+    try{
+        if(session.getAttribute("idx") == null){
+            isLoginIn = false;
+            throw new Exception("접근 권한이 없습니다.");
+        }
 
-    query.setInt(1, scheduleIdxValue);
-    query.setInt(2, Integer.parseInt(String.valueOf(session.getAttribute("idx"))));
-    query.executeUpdate();
+        if(session.getAttribute("idx") != null){
+            Class.forName("com.mysql.jdbc.Driver"); 
+            Connection conn = DriverManager.getConnection("jdbc:mysql://localhost:3306/scheduler", "suin", "suin"); 
+
+            String sql = "DELETE FROM schedule WHERE idx=? AND user_idx=?";
+            PreparedStatement query = conn.prepareStatement(sql); 
+
+            query.setInt(1, scheduleIdxValue);
+            query.setInt(2, Integer.parseInt(String.valueOf(session.getAttribute("idx"))));
+            query.executeUpdate();
+        }
+    }catch(Exception e){
+        errMessage = e.getMessage();
+    }
+
 %>
 
 <head>
@@ -30,6 +45,14 @@
 <body>
     <script>
         var dateValue = <%=dateValue%>;
-        location.href = `../page/modal.jsp?date=\${dateValue}`;
+        var errMessage = "<%=errMessage%>";
+        var isLoginIn = <%=isLoginIn%>;
+
+        if(!isLoginIn){
+            alert(errMessage);
+            location.href = "../page/login.html";
+        }else{
+            location.href = `../page/modal.jsp?date=\${dateValue}`;
+        }
     </script>
 </body>
